Separate globe atmosphere rendering from sky (#11344)
* Separate globe atmosphere rendering from sky

* Render sky layer during the transition between mercator and globe
